[D3D11] Removed D3D11NativeTexture union.
C++ unions must only read the member that was last written to, which is disregarded with D3D11NativeTexture.
It also uses ComPtr as members, which is not suitable for unions as they should only use plain-old-data members.
All uses of the subtypes (ID3D11Texture1D/2D/3D) are cast dynamically using ComPtr::As<T>() when needed.