[SwapChain] Make default surface visible at creation time with an auto-generated title when creating a swap-chain.
The rationale of making the surface invisible when creating a swap-chain was to prevent the window to be visible before it has a valid title.
The client programmer was responsible for changing the title and then making the window (or canvas on mobile) visible.
This is not an issue when a custom surface is passed into the CreateSwapChain() function, but the default surface never had an initial title.

With this change, the default surface is always visible at creation time and an auto-generated title is used to initialzie the window/canvas.
If this behavior is undesired, the client programmer can continue to use a custom surface but the default surface is supposed to make things easier.