[GL] Deprecated Surface::ResetPixelFormat() by selecting multi-sample pixel format via proxy window for Win32 platform.
Surface::ResetPixelFormat() was solely used to re-create the internal window of a Win32 LLGL::Window object when a multi-sample pixel format is selected for an OpenGL context.
In order to load the necessary GL extension procedure wglChoosePixelFormatARB(), a valid GL context must be active. After selecting a format, the multi-sampled GL context can be created.
A Win32 window, however, cannot change its pixel format once its selected, so previously ResetPixelFormat() would re-create the native window on Win32 platform, while all other platforms only had dummy implementations.
This change uses a proxy window that is invisible to select a multi-sample pixel format when a WGL context is created before the format is selected for the actual window provided for the swap-chain.