[Android] Fixed event handling of pause/resume Android app.
When an Android app is paused, its ANativeWindow is destroyed (APP_CMD_TERM_WINDOW) and the dependent EGLSurface must be destroyed as well.
When the app resumes, the ANativeWindow is re-initialized (APP_CMD_INIT_WINDOW) and the depednent EGLSurface must be re-created.
This is handled via a Canvas event listener for the AndroidGLSwapChainContext that responds to the new callbacks OnInit and OnDestroy.

Additional notes:
- Deprecated OnQuit and PostQuit functions of Canvas event handlers; This was only intended for desktop platforms when a window close button is pressed.
- Added OnInit and OnDestroy functions to Canvas::EventListener interface.
- Moved Android input event handling from AndroidCanvas.cpp file to new AndroidInputEventHandler.cpp/.h files.