[Examples] Started to replace static members in examples with local class members.
The way Android apps load native .so libs causes static members to be initialized only once.
After the app is closed, even when the main entry point has exited, the next re-launch of the app
will retain the previous values of those static members because JNI does not necessarily re-load the shared library.
This makes the app partially look as if it never re-loaded.