[SwapChain] Added IsPresentable() to SwapChain interface.
This new function allows to detect when the native surface of a swap-chain is temporarily unavailable.
This occurs when a mobile app is paused and the native surface is destroyed, while it is later re-initialized when the app resumes.
Also moved all SwapChain interface declarations to Backend/SwapChain.inl header as for other interfaces with more than one or two functions.