[JIT] Removed JIT compiler.
The JIT compiler was never well supported, in fact, it only worked somewhat acceptable on Intel Macs and was mostly an experimental project within LLGL.
This change removes the JIT compiler entirely from the project and non-native deferred command buffers are now only emulated.

The performance gain through a JIT compiled command buffer was questionable and never tested to begin with;
Most time is spent in the GL calls themselves, not the switch-case statement in the emulated equivalent.

This is a non-breaking change as it was only configurable through CMake. The removed option LLGL_ENABLE_JIT_COMPILER will now be ignored.