[GLX] Fixed SIGSEGV exception in GLXContext teardown.
When the X11 display connection is closed, any shutdown callback that libGL.so might have registered will be called.
This causes a SIGSEGV exception if the GL backend has previously been unloaded, even though the GLX context was properly destroyed.
To prevent this, the LLGL::LinuxSharedX11Display singleton retains a reference to libGL.so if it was used by the GL backend
and we only unload this library *after* the call to XCloseDisplay().