[Examples] Refactored Tessellation example.
The new tessellation example makes better use of shaders and textures to illustrate a more practial use of hardware tessellation.
It renders an ordinary cube with a color, normal, specular (phong-shading), and displacement map.
The cube model has a single subdivions as pre-computed model to increase the maximum tessellation factor.

TODO: Metal shaders need to be updated.