[Samplers] Added "CombinedTextureSamplerDescriptor" to simplify sampler binding in OpenGL.
This is a larger refactor of the PSO layout for the GL backend to simplify the binding model of combined texture-samplers.
Previously, each texture binding needed the same number of sampler bindings at the same binding slots.
This change unifies the GL backend further with the other backends by adding a separate list in the PSO layout of combined samplers.
A detailed example is provided in the documentation for the new public structure "CombinedTextureSamplerDescriptor".

This also adds a new syntax to the Parse() function, e.g. "sampler<tex, smpl>(tex_smpl@1)" specifies such a combined texture-sampler.

Added new unit test "CombinedTexSamplers".