[API] Add GetNativeHandle() to public Resource interface (#142).
This adds another version of "GetNativeHandle()" functions to the public interface for resources.
While LLGL is intended to highly encapsulate lower level resources, this is one of the few functions to get access to the internal resource types.
This is mean to improve interoperability between LLGL and other graphics libraries such as ImGui.

The test "NativeHandle" was also added to Testbed (NOTE: this test is incomplete and needs more test cases).