[Misc] Renamed Throw.h to Trap.h and made Trap() publicly available.
This builds upon the new way to better handle exceptions in the public interface (671c74ed).
Instead of only calling abort() when exceptions are disabled, this change makes use of the
already implemented Trap() function that provides a full callstack and other functionality.
Since LLGL does not handle but only throws exceptions, we accept that this only provides a
handful of predefined exceptions for the few cases they are used in the public header files.