[D3D11] Fix crash inside WARP renderer when intermediate raw buffer is smaller than 16 bytes.
This is a workaround for an issue that appears to be a bug inside Microsoft's WARP renderer,
which makes the CIS build fail for the test case 'Copy texture "tex{Cube[2],RgbaUb,2wh}" (mip: 1, layer: 0) to buffer'.
The WARP renderer crashes with a read-access violation when executing the built-in compute shader
to copy a buffer of size 4 bytes to a RWTexture2DArray of dimension 1x1x1 (with R32UInt format).
This is fixed by clamping the interemdiate buffer size to 16 bytes.