[CMake] Also set LIBRARY_OUTPUT_DIRECTORY target property.
This is fix to the previous change of using the RUNTIME_OUTPUT_DIRECTORY target property in CMakeLists.txt
MSVC does not have to rely on LIBRARY_OUTPUT_DIRECTORY for shared libraries, i.e. add_library() target, but Linux does.
This fixes the missing libLLGL_*.so libraries in the output directory on Linux.