[Mac] Fixed Display::GetCursorPosition() for MacOS.
This was using the completely wrong function.
The global cursor position can be queried via [NSEvent mouseLocation].
Adjust the position in the same way as in MacOSWindow to return the position
in LLGL's coordinate system with (0,0) being the top-left corner of the screen that windows can reach,
i.e. the area of Mac's menu bar is in the negative for the Y-coordinate.