[VK] Always enable VK_IMAGE_USAGE_TRANSFER_SRC_BIT in buffer and textures.
This must always be enabled since LLGL does not require BindFlags::CopySrc flag for ReadBuffer/ReadTexture functions.
In the future, the API should be changed to require CPUAccessFlags::Read for these functions to be used.
This will require some refactoring regarding the secondary buffers that LLGL currently allocates when CPU access is requested.