[D3D12] Started with refactoring of Map/UnmapBuffer interface in D3D12 backend.
The MapBuffer/UnmapBuffer functions in the RenderSystem interface don't provide much benefit over ReadBuffer/WriteBuffer.
This refactoring initiative attempts to provide a better interface to make effective use of mapping GPU buffers into CPU memory space.
The idea is to move the mapping functions into the Buffer interface to make it clear they can be used during command encoding.
The Buffer::Map() function includes an extensive example of how to use it for large buffer updates as an alternative to UpdateBuffer(),
which is limited to 2^16 bytes. This example also illustrates the first beneficial use of the Fence interface to synchronize memory updates.

TODO: Since D3D12 requires a strict separation between readback and upload buffers,
      the D3D12 backend must keep them in sync if a buffer is created with both read and write access.