[API] Lifted restriction on UpdateBuffer() size (fixes #156).
This is lifting the size restriction that limited UpdateBuffer() updates to 2^16-1 bytes by changing the 'dataSize' parameter from uint16 to uint64.
This restriction only existed because of the limited size for vkCmdUpdateBuffer() in the Vulkan backend.
The Vulkan backend now provides an alternative path for larger updates that uses the same kind of staging buffer pool as in the D3D12 and Metal backends.
Also updated BufferUpdate unit test to include a buffer update of size 2^18.