[Metal] Ensure MTLRenderCommandEncoder is bound when MTDirectCommandBuffer::GetNativeHandle() is called.
Due to delayed commands in MTCommandContext, the GetNativeHandle() implementation of the immediate command buffer
in the Metal backend needs to bind the render command encoder when its inside a render pass.
Relying on the command encoder already being bound is not sufficient since BeginRenderPass()
does not actually start the command encoder.