[VK] Added TexelBuffer binding flag as a workaround for Vulkan resource heaps.
The idea was to let the Vulkan backend determine automatically whether texel buffers or SSBOs are bound in a PSO.
This is not trivial for resource heaps because the VkDescriptorSetLayout has to know the exact descriptor type at creation time.
For that reason, this commit is a workaround which requires the client to specify this information at the public interface.
The new binding flag BindFlags::TexelBuffer should be removed once a better solution has been found.